Add files via upload
This project was done in pivot table and data was taken from kaggale.com. In this Project our problem statement was to find the people living in Kolkota reduce their rent by 10% by monitoring rent price with respect with Furnishing Status, Area Type, Built Area, Point of Contact, Number of Bathroom by the end of the year 2022.